refactor: organize route fallback skeletons
chore
